+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* set all extra header values in `Base::parse_extra_headers()`
 
 #### 7.4.2 / 2017-10-25
 * added check to see if wp-cron is updating and if not send and error message